Add files via upload
Rework and expansion of the MxKoO conky... many add-ons and improvements including Battery state data that changes color based on Batt %, multiple columns of data in one line, detailed Network info, and more... I did not make significant color changes (yet) but expect to publish a similar version that is monochrome or at least milder/different color schemes.

This Conky is setup for Linux Mint (or similar) but can be adapted (of course) for most any system. Remember to change the references to your system as they may not function without it.  Enjoy!...